Extracted Frames from the Canvas
The Canvas now only references a Frame object which holds the actual colour grid, with the canvas supplying the addition methods needed to modify the frame.

The canvas now displays the frames in the animator, instead of the animator getting the grid from the canvas. This is great because now we can use this frame when displaying it in places other than the canvas; preview, animator etc