Refactored Canvas to use Events
What happens when you click the canvas is now determined by an event.

Also...
I found the need for a clear method on the events to remove all subscribed methods in one go, so I dove into the source code and added one. It does go against what events should technically be, but for this simple project it will help me a lot